requirements.txt
# Core ML
torch==2.9.1
torchaudio==2.9.0
numpy==2.0.2
pandas==2.2.3
scikit-learn==1.5.2

# Audio Processing
librosa==0.10.2
pydub==0.25.1
audiomentations==0.36.0

# Visualization
matplotlib==3.9.2
seaborn==0.13.2
umap-learn==0.5.6

# NLP for lyrics
sentence-transformers==3.2.1
transformers==4.46.2
langdetect==1.0.9

# Utilities
tqdm==4.66.5
pyyaml==6.0.2
jupyter==1.0.0
kaggle==1.6.17